\name{check_convergence}
\alias{check_convergence}

\title{Check Convergence of Monte Carlo Samples}

\description{Function to check convergence of mean, median and 95\% uncertainty interval of global DALY estimate.}

\usage{check_convergence(x, n = 100L, denom = 1e5)}

\arguments{
  \item{x}{object generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
  \item{n}{convergence of last \code{n} samples is evaluated}
  \item{denom}{denominator used in rates}
}

\value{Generates plot of running mean, median and 95\% uncertainty interval; returns average and maximum absolute deviation of mean, median and 95\% uncertainty interval.}

\keyword{report}
