\name{DALY-helpers}
\alias{DALY-helpers}
\alias{get_contrib}
\alias{mean_age}
\alias{get_local_LE}

\title{DALY Helper Functions}

\description{Set of exported helper functions.}

\usage{
get_contrib(x)
mean_age()
get_local_LE(country)
}

\arguments{
  \item{x}{object generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
  \item{country}{country name or integer identification number}
}

\details{\code{mean_age} reads an age distribution from the clipboard, and calculates through linear interpolation the corresponding mean ages and WHO life expectancies.}

\value{\code{get_contrib} returns vector of node contributions (ie, YLD or YLL); \code{mean_age} returns a list of mean ages and life expectancies; \code{get_local_LE} returns the age-specific life expectancies for the concerned country.}

\examples{
## First country in dataset is Afghanistan
get_local_LE(1)
get_local_LE("Afghanistan")
}

\keyword{DALY}
