\name{DALY-summarize}
\alias{DALY-summarize}
\alias{summarize}
\alias{summarize_all}
\alias{load_results}
\alias{get_total}
\alias{get_global}
\alias{get_samples}
\alias{get_all_samples}
\alias{get_samples_per_outcome}

\title{Load, Merge, and Summarize Results by (Sub)Region}

\description{
Functions to load, merge and summarize the number of incident cases, deaths, YLDs, YLLs or DALY of \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}} objects.
}

\usage{
## Get summary for 1 agent
summarize(x, what, rate = FALSE, age = c("all", "<5", "5+"), 
          denom = 1e+05, names = "all")

## Get summary for all agents 
summarize_all(agents, what = "daly", rate = F, by = "region",
              age = "all", fb = F)

## Load (fb) results per (sub)region
load_results(agent, by, fb)

## Get summary per (sub)region for 1 agent
get_total(agent, what = "daly", rate = FALSE,
          by = c("subregion", "region"), age = c("all", "<5", "5+"), fb = FALSE)

##
get_global(x, what, by_region = FALSE, denom = 1e+05, rate = FALSE, 
           names = "all", age = "all")

## Get DALY samples by (sub)region for 1 agent
get_samples(agent, what = c("daly", "yld", "yll", "incidence", "cases", "deaths"),
            by = c("subregion", "region"), fb = FALSE, ...)

## Get DALY samples by (sub)region for >1 agent
get_all_samples(agents, what = c("daly", "yld", "yll", "incidence", "cases", "deaths"),
                by = c("subregion", "region"), fb = FALSE, ...)

## Get all samples, per outcome
get_samples_per_outcome(agent, dir, date, denom = 1e+05,
                        by = c("subregion", "region"), outcomes)
}

\arguments{
  \item{x}{object generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
  \item{what}{parameter to be summarized}
  \item{rate}{logical flag; should values be summarized per 100,000?}
  \item{age}{age group for which summary is made}
  \item{denom}{denominator used in rates}
  \item{names}{node name(s) for which summary is made}
  
  \item{agent}{vector of 4 elements: agent name, directory, date total burden, date foodborne burden}
  \item{agents}{list of \code{agent} vectors}
  \item{by}{which set of results to use: \code{c("subregion", "region")}}
  \item{fb}{logical flag; should foodborne results be used?}
  \item{by_region}{logical flag; should results be presented at (sub)regional level, or should they be aggregated at global level?}
  \item{dir}{agent directory}
  \item{date}{date burden estimation}
  \item{outcomes}{vector of outcomes to be used}
  
  \item{\dots}{other arguments to be passed to underlying functions}
}

\value{
The \code{summarize} functions return a data frame with the mean, median, 95\% uncertainty interval, and standard deviation of the specified value(s) per (sub)region and globally. The \code{get} functions return a list of samples according to the arguments specified.
}

\seealso{\code{\link{DALY_map}}, \code{\link{FERG_report}}}

\keyword{report}
