\name{DALY_map}
\alias{DALY_map}

\title{Global Burden of Disease Map}

\description{This function generates a worldmap of the burden of disease by region or subregion.}

\usage{
DALY_map(x, agent, cat = "pretty", numCats = 7,
         palette = brewer.pal(numCats, "Reds"), 
         what = c("daly", "sd", "inc", "mrt"),
         save = TRUE, type = c("png", "tiff"),
         legend_title, legend_digits = 3)
}

\arguments{
  \item{x}{object generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
  \item{agent}{name of the agent}
  \item{cat}{name of the categorization scheme}
  \item{numCats}{number of colour categories}
  \item{palette}{name of the colour palette}
  \item{what}{number of colour categories}
  \item{save}{logical flag; should the image be saved?}
  \item{type}{logical flag; save the image as a \code{png} or \code{tiff} file?}
  \item{legend_title}{title of the legend}
  \item{legend_digits}{number of digits in the legend labels}
}

\value{As side effect, a \code{png} or \code{tiff} file is created if \code{save = TRUE}.}

\seealso{\code{\link{DALY-plots}}, \code{\link{DALY-summarize}}, \code{\link{FERG_report}}}

\keyword{report}
