\name{FERG_report}
\alias{FERG_report}

\title{Generate FERG PDF Report}

\description{
This function generates a PDF report of the global burden of foodborne disease. The report consists of a numerical and graphical summary of the \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}} object.
}

\usage{
FERG_report(DALY, agent, agent_full, dm_file, today, by_age = FALSE, 
            denom = 1e+05, digits_net = 0, digits_rate = 3, scale = 0.8, 
            add = NULL, fb = FALSE, exp = NULL, DALY_total = NULL)
}

\arguments{
  \item{DALY}{object generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
  \item{agent}{agent name, abbreviated}
  \item{agent_full}{agent name, full}
  \item{dm_file}{link to PDF file of disease model}
  \item{today}{today's date}
  \item{by_age}{logical flag; should results be split up by age group?}
  
  \item{denom}{denominator used in rates}
  \item{digits_net}{number of significant digits for net values}
  \item{digits_rate}{number of significant digits for rates}
  \item{scale}{size of scalebox for tables}
  
  \item{add}{list of node names that need to be merged}
  \item{fb}{logical flag; should foodborne results be presented?}
  \item{exp}{source attribution > exposure samples}
  \item{DALY_total}{source attribution > object of total DALYs generated by \code{\link{getDALY_per_region}} or \code{\link{getDALY_per_subregion}}}
}

\value{As side effect, a PDF report is generated.}

\seealso{\code{\link{summarize}}, \code{\link{DALY_map}}}

\keyword{report}
