\name{source-attribution-functions}
\alias{source-attribution-functions}
\alias{generate_samples}
\alias{normalize_by_food}
\alias{apply_prop_fb}

\title{Main Source Attribution Functions}

\description{Functions to generate normalized samples; combine exposure and food item samples; and generate foodborne burden estimates.}

\usage{
## Generate normalized samples
generate_samples(n, file, cols, names)

## Normalize samples by proportion foodborne
normalize_by_food(food, exp)

## Apply proportion foodborne to burden estimates
apply_prop_fb(agent, date_DALY, exp)
}

\arguments{
  \item{n}{number of samples}
  \item{file}{\code{*.dis} file generated by \pkg{EXCALIBUR}}
  \item{cols}{\code{*.dis} file columns to use}
  \item{names}{names of exposure routes or food items}
  \item{food}{list of food item samples}
  \item{exp}{list of exposure route samples}
  \item{agent}{agent name, abbreviated}
  \item{date_DALY}{date of DALY calculation}
}

\value{\code{generate_samples} and \code{normalize_by_food} return a data frame of normalized samples; \code{apply_prop_fb} creates \sQuote{\code{DALYrun}} objects per country.}

\seealso{\code{\link{source-attribution-report}}}

\keyword{SA}
